Added option to cache FBProfile images beyond their lifetime
This is very helpful when using the class in a UITableView, where
images are replaced when the UITableView scrolls. Instead of the cached
image having a ‘weak’ retention, if the class is initialised by
correctly, the cached images will have a ‘strong’ retention. Every 30
seconds, the cache will be reduced to the maximum number of images that
was passed on initialisation.